Minor: Move `project_schema` to `datafusion_common`
